+++ explainaboard/processors/language_modeling.py
@@ -17,7 +17,7 @@ from explainaboard.metric import LogProbConfig, MetricConfig, MetricStats
 from explainaboard.processors.processor import Processor
 from explainaboard.processors.processor_registry import register_processor
 from explainaboard.utils import bucketing
-from explainaboard.utils.analysis import cap_feature
+from explainaboard.utils.feature_funcs import cap_feature
 from explainaboard.utils.logging import progress
 from explainaboard.utils.typing_utils import unwrap
